Update README.md added go get
The command `go get github.com/cucumber/godogsteps` is missing and this will lead to ohter errors later. So if you add this line then the rest should be as expected. Although im do not know if `sum.go` is creating after `go get ...` or after the next command `go test`